Deep scan: Fix encoding and verify plug-in-out compliance
- Removed BOM from 6 files (CorridorFillSystem, CorridorFlowSystem, MazeObjectSpawner, MazeDoorSpawner, MazeMarkerSpawner, DeadEndCorridorSystem)
- Converted CRLF to LF line endings on all 6 files
- All files now UTF-8 without BOM, Unix LF line endings
- Verified EventHandler plug-in-out compliance across active systems
- Build successful, no errors or warnings

+++ Assets/Scripts/Core/06_Maze/ProceduralLevelGenerator.cs
@@ -277,7 +277,7 @@ namespace Code.Lavos.Core
 			{
 				for (int z = 0; z < mazeData.Height; z++)
 				{
-					uint cell = mazeData.GetCell(x, z);
+					uint cell = (uint)mazeData.GetCell(x, z);
 					Vector3 position = new Vector3(x * mazeData.Config.CellSize, 0f, z * mazeData.Config.CellSize);
 
 					// Place pillars
